Added missing NetworkModule.kt and inject retrofit client
