--- app/src/main/java/com/example/ncov19traking/api/NCoVApiClient.kt
@@ -1,23 +1,0 @@
-package com.example.ncov19traking.api
-
-import okhttp3.OkHttpClient
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-object NCoVApiClient {
-
-    val nCoVApi by lazy {
-        Retrofit.Builder()
-            .baseUrl("https://corona.lmao.ninja/")
-            .client(httpInterceptor)
-            .addConverterFactory(GsonConverterFactory.create())
-            .build()
-            .create(NCoVApi::class.java)
-    }
-
-    private val httpInterceptor = OkHttpClient.Builder().addInterceptor { chain ->
-        val request = chain.request()
-        val response = chain.proceed(request)
-        response
-    }.build()
-}
