+++ app/src/main/java/com/example/ncov19traking/data/NCoVRepository.kt
@@ -15,12 +15,13 @@ class NCoVRepository @Inject constructor(
     private val nCoVDao: NCoVDao,
     private val countryDao: CountryDao,
     private val globalHistoricalDao: GlobalHistoricalDao,
+    private val nCoVApi: NCoVApi,
     private var errorResponse: MutableLiveData<ErrorBody>
 ) {
 
     suspend fun getAllCases(): NCoVInfo {
         try {
-            when (val response = ApiResponse.create(NCoVApiClient.nCoVApi.getGeneralNumbers())) {
+            when (val response = ApiResponse.create(nCoVApi.getGeneralNumbers())) {
                 is ApiSuccessResponse -> nCoVDao.save(response.data)
                 is ApiSuccessEmptyResponse -> errorResponse.postValue(ErrorBody(message = "Empty body"))
                 is ApiErrorResponse -> errorResponse.postValue(
@@ -45,7 +46,7 @@ class NCoVRepository @Inject constructor(
     suspend fun getAllYesterdayCases(): NCoVInfoYesterday {
         try {
             when (val response =
-                ApiResponse.create(NCoVApiClient.nCoVApi.getYesterdayGeneralNumbers())) {
+                ApiResponse.create(nCoVApi.getYesterdayGeneralNumbers())) {
                 is ApiSuccessResponse -> nCoVDao.saveYesterday(response.data)
                 is ApiSuccessEmptyResponse -> errorResponse
                 is ApiErrorResponse -> Log.d("apiError", "${response.code}: ${response.message}")
@@ -59,7 +60,7 @@ class NCoVRepository @Inject constructor(
 
     suspend fun getAllCountries(): Array<NumbersByCountry>? {
         try {
-            when (val response = ApiResponse.create(NCoVApiClient.nCoVApi.getNumbersByCountry())) {
+            when (val response = ApiResponse.create(nCoVApi.getNumbersByCountry())) {
                 is ApiSuccessResponse -> countryDao.save(response.data)
                 is ApiSuccessEmptyResponse -> errorResponse.postValue(ErrorBody(message = "Empty body"))
                 is ApiErrorResponse -> errorResponse.postValue(
@@ -78,12 +79,12 @@ class NCoVRepository @Inject constructor(
         return countryDao.load()
     }
 
-    suspend fun getHistoricalCountryData() = NCoVApiClient.nCoVApi.getHistoricalDataByCountry()
+    suspend fun getHistoricalCountryData() = nCoVApi.getHistoricalDataByCountry()
 
     suspend fun getAllHistoricalDataCases(): Timeline? {
         try {
             when (val response =
-                ApiResponse.create(NCoVApiClient.nCoVApi.getAllHistoricalData())) {
+                ApiResponse.create(nCoVApi.getAllHistoricalData())) {
                 is ApiSuccessResponse -> globalHistoricalDao.save(response.data)
                 is ApiSuccessEmptyResponse -> errorResponse.postValue(ErrorBody(message = "Empty body"))
                 is ApiErrorResponse -> errorResponse.postValue(
